+++ components/main/index.js
@@ -1,0 +1,23 @@
+import React, { Component } from 'react';
+import { View, Text } from 'react-native';
+
+import styles from './styles';
+
+export default class TweetstormRN extends Component {
+  render() {
+    return (
+      <View style={styles.container}>
+        <Text style={styles.welcome}>
+          Welcome to React Native!
+        </Text>
+        <Text style={styles.instructions}>
+          To get started, edit index.android.js
+        </Text>
+        <Text style={styles.instructions}>
+          Double tap R on your keyboard to reload,{'\n'}
+          Shake or press menu button for dev menu
+        </Text>
+      </View>
+    );
+  }
+}
